Move `commit_is_ancestor_of_head` to history mixin
